requirements.txt
openai
python-dotenv
fastapi
uvicorn[standard]
langchain
langchain-core
langchain-openai